Log starboard message events, fixes
1) When a starboard message is added or removed, it will be logged in rina's log channel (named 'vcLog' inconveniently)
2) If you react to the original message with the starboard message multiple times after it reaches the minimum requirement but before rina can add her reaction, it would send duplicate starboard messages. This is fixed.
3) Fix cool debug return cursor thingy in Rina startup message

+++ cmdg_Starboard.py
@@ -27,6 +27,7 @@ class Starboard(commands.Cog):
             original_message = await ch.fetch_message(message_id)
         except discord.NotFound:
             # if original message removed, remove starboard message
+            await logMsg(star_channel.guild, f"{starboard_emoji} :x: Starboard message {star_message.id} was removed (from {message_id}) (original message could not be found)")
             await star_message.delete()
             return
 
@@ -56,10 +57,11 @@ class Starboard(commands.Cog):
                 if reaction.me:
                     star_stat += 1
 
-        #todo
         #if more x'es than stars, and more than 15 reactions, remove message
         if star_stat < 0 and reactionTotal > 10:
-           await star_message.delete()
+            await logMsg(star_channel.guild, f"{starboard_emoji} :x: Starboard message {star_message.id} was removed (from {message_id}) (too many downvotes! Score: {star_stat}, Votes: {reactionTotal})")
+            await star_message.delete()
+            return
 
         # update message to new star value
         parts = star_message.content.split("**")
@@ -110,7 +112,7 @@ class Starboard(commands.Cog):
                                 await self.updateStat(star_message)
                                 return
                     return
-                elif reaction.count > star_minimum:
+                elif reaction.count = star_minimum:
                     if message.author == self.client.user:
                         #can't starboard Rina's message
                         return
@@ -139,7 +141,7 @@ class Starboard(commands.Cog):
                             embed = embed,
                             allowed_mentions = discord.AllowedMentions.none(),
                         )
-
+                    await logMsg(star_channel.guild, f"{starboard_emoji} Starboard message {msg.jump_url} was created from {message.jump_url}")
                     # add initial star reaction to starboarded message, and new starboard msg
                     await message.add_reaction(starboard_emoji)
                     await msg.add_reaction(starboard_emoji)
@@ -197,12 +199,18 @@ class Starboard(commands.Cog):
             raise KeyError("Not enough data is configured to .. check starboard for a message matching the deleted message's ID, because idk what channel i need to look in! Please fix this with `/editguildinfo`!")
         star_channel = self.client.get_channel(_star_channel)
 
-        # check if this message's is in the starboard. If so, delete it
-        async for star_message in star_channel.history(limit=200):
-            for embed in star_message.embeds:
-                if embed.footer.text == str(message_payload.message_id):
-                    await star_message.delete()
-                    return
+        if message_payload.channel_id == star_channel.id:
+            # check if the deleted message is a starboard message; if so, log it at starboard message deletion
+            await logMsg(star_channel.guild, f"{starboard_emoji} :x: Starboard message {star_message.id} was removed (from {message_payload.message_id}) (Starboard message was deleted manually)")
+            return
+        elif message_payload.channel_id != star_channel.id:
+            # check if this message's is in the starboard. If so, delete it
+            async for star_message in star_channel.history(limit=300):
+                for embed in star_message.embeds:
+                    if embed.footer.text == str(message_payload.message_id):
+                        await logMsg(star_channel.guild, f"{starboard_emoji} :x: Starboard message {star_message.id} was removed (from {message_payload.message_id}) (original message was removed (this starboard message's linked id matched the removed message's))")
+                        await star_message.delete()
+                        return
 
 async def setup(client):
     await client.add_cog(Starboard(client))
